more casting because I can't code

+++ Scanner/Program.cs
@@ -75,7 +75,7 @@ namespace ByteScanner
                 foreach (var result in patternResult.Where(x => x.Found))
                 {
                     Console.WriteLine("    Got hit on address: 0x{0:X}", ((long)result.Offset + (long)module.BaseAddress));
-                    Console.WriteLine("    Base address: 0x{0:X} on offset 0x{1:X}", module.BaseAddress, (ulong)result.Offset);
+                    Console.WriteLine("    Base address: 0x{0:X} on offset 0x{1:X}", (long)module.BaseAddress, (ulong)result.Offset);
                 }
             }
             else {
